requirements_new.txt

# Telegram Bot
telethon>=1.35.0

# HTTP Requests and Web Scraping
aiohttp>=3.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Analysis and Technical Indicators
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0

# Database
aiosqlite>=0.19.0

# Configuration and Environment
python-dotenv>=1.0.0

# Async and Concurrency
asyncio-throttle>=1.0.2

# Utilities
schedule>=1.2.0
